+++ src/electron/window/onenote/load.js
@@ -26,6 +26,7 @@ const langTranslations = {
     'es-ES': require('../../../translation/es-ES'),
     'fr-FR': require('../../../translation/fr-FR'),
     'nl-NL': require('../../../translation/nl-NL'),
+    'it-IT': require('../../../translation/it-IT'),
 
 }