+++ changelog.md
@@ -7,6 +7,9 @@
 
 ## Change log
 
+### v2021.4.171
+* FEATURE: Added Italian translation.
+
 ### v2021.4.169
 * FEATURE: Enable dark mode (with quirks) using invert all colors. Not perfect, but it works and is bettter for you eyes.
 
@@ -258,7 +261,7 @@ All my domains ([patrikx3.com](https://patrikx3.com) and [corifeus.com](https://
 
 ---
 
-[**P3X-ONENOTE**](https://corifeus.com/onenote) Build v2021.4.169
+[**P3X-ONENOTE**](https://corifeus.com/onenote) Build v2021.4.171
 
 [![Donate for Corifeus / P3X](https://img.shields.io/badge/Donate-Corifeus-003087.svg)](https://www.paypal.com/cgi-bin/webscr?cmd=_s-xclick&hosted_button_id=QZVM4V6HVZJW6)  [![Contact Corifeus / P3X](https://img.shields.io/badge/Contact-P3X-ff9900.svg)](https://www.patrikx3.com/en/front/contact) [![Like Corifeus @ Facebook](https://img.shields.io/badge/LIKE-Corifeus-3b5998.svg)](https://www.facebook.com/corifeus.software)